+++ examples/bitmap_operation/README.md
@@ -5,7 +5,7 @@ Open example_bitmap_operation.vcxproj file in this directory to create solution 
 - g++    
 In this directory you need to type/paste this text in terminal:    
 	```bash
-	g++ -std=c++11 -Wall example_bitmap_operation.cpp ../../src/image_function_helper.cpp ../../src/image_function.cpp ../../src/FileOperation/bitmap.cpp -o application
+	g++ -std=c++11 -Wall example_bitmap_operation.cpp ../../src/image_function_helper.cpp ../../src/image_function.cpp ../../src/file/bmp_image.cpp -o application
 	```
 
 - make
