+++ examples/bitmap_operation/example_bitmap_operation.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include "../../src/image_buffer.h"
 #include "../../src/image_function.h"
-#include "../../src/FileOperation/bitmap.h"
+#include "../../src/file/bmp_image.h"
 
 void method1();
 void method2();
