+++ examples/bitmap_operation/example_bitmap_operation.vcxproj
@@ -145,11 +145,11 @@
   <ItemGroup>
     <ClCompile Include="..\..\src\image_function_helper.cpp" />
     <ClCompile Include="example_bitmap_operation.cpp" />
-    <ClCompile Include="..\..\src\FileOperation\bitmap.cpp" />
+    <ClCompile Include="..\..\src\file\bmp_image.cpp" />
     <ClCompile Include="..\..\src\image_function.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="..\..\src\FileOperation\bitmap.h" />
+    <ClInclude Include="..\..\src\file\bmp_image.h" />
     <ClInclude Include="..\..\src\image_buffer.h" />
     <ClInclude Include="..\..\src\image_exception.h" />
     <ClInclude Include="..\..\src\image_function.h" />
