+++ examples/bitmap_operation/makefile
@@ -3,7 +3,7 @@
 ##
 CXXFLAGS += -std=c++11 -Wall -Wextra -Wstrict-aliasing -Wpedantic -Wconversion -O2 -march=native
 
-example_bitmap_operation : ../../src/image_function_helper.cpp ../../src/image_function.cpp ../../src/FileOperation/bitmap.cpp
+example_bitmap_operation : ../../src/image_function_helper.cpp ../../src/image_function.cpp ../../src/file/bmp_image.cpp
 
 .PHONY: clean
 clean:
