+++ examples/blob_detection/CMakeLists.txt
@@ -3,6 +3,6 @@ add_executable(example_blob_detection
     ${LIB_DIR}/blob_detection.cpp
     ${LIB_DIR}/image_function_helper.cpp
     ${LIB_DIR}/image_function.cpp
-    ${LIB_DIR}/FileOperation/bitmap.cpp)
+    ${LIB_DIR}/file/bmp_image.cpp)
 target_link_libraries(example_blob_detection
     PRIVATE example_features_options example_features_warnings)
