+++ examples/blob_detection/README.md
@@ -5,7 +5,7 @@ Open example_blob_detection.vcxproj file in this directory to create solution fo
 - g++    
 In this directory you need to type/paste this text in terminal:    
 	```bash
-	g++ -std=c++11 -Wall example_blob_detection.cpp ../../src/math/math_base.cpp ../../src/image_function_helper.cpp ../../src/image_function.cpp ../../src/blob_detection.cpp ../../src/FileOperation/bitmap.cpp -o application
+	g++ -std=c++11 -Wall example_blob_detection.cpp ../../src/math/math_base.cpp ../../src/image_function_helper.cpp ../../src/image_function.cpp ../../src/blob_detection.cpp ../../src/file/bmp_image.cpp -o application
 	```
 
 - make
