+++ examples/blob_detection/example_blob_detection.cpp
@@ -3,7 +3,7 @@
 #include "../../src/blob_detection.h"
 #include "../../src/image_buffer.h"
 #include "../../src/image_function.h"
-#include "../../src/FileOperation/bitmap.h"
+#include "../../src/file/bmp_image.h"
 
 void example1();
 void example2();
