+++ examples/blob_detection/example_blob_detection.vcxproj
@@ -147,12 +147,12 @@
     <ClCompile Include="..\..\src\math\math_base.cpp" />
     <ClCompile Include="example_blob_detection.cpp" />
     <ClCompile Include="..\..\src\blob_detection.cpp" />
-    <ClCompile Include="..\..\src\FileOperation\bitmap.cpp" />
+    <ClCompile Include="..\..\src\file\bmp_image.cpp" />
     <ClCompile Include="..\..\src\image_function.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\src\blob_detection.h" />
-    <ClInclude Include="..\..\src\FileOperation\bitmap.h" />
+    <ClInclude Include="..\..\src\file\bmp_image.h" />
     <ClInclude Include="..\..\src\image_buffer.h" />
     <ClInclude Include="..\..\src\image_exception.h" />
     <ClInclude Include="..\..\src\image_function.h" />
