+++ examples/cuda/image_function/example_image_function.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include "../../../src/image_buffer.h"
 #include "../../../src/image_function.h"
-#include "../../../src/FileOperation/bitmap.h"
+#include "../../../src/file/bmp_image.h"
 #include "../../../src/cuda/cuda_device.cuh"
 #include "../../../src/cuda/cuda_helper.cuh"
 #include "../../../src/cuda/image_buffer_cuda.cuh"
