+++ examples/filters/CMakeLists.txt
@@ -3,6 +3,6 @@ add_executable(example_filters
     ${LIB_DIR}/image_function_helper.cpp
     ${LIB_DIR}/image_function.cpp
     ${LIB_DIR}/filtering.cpp
-    ${LIB_DIR}/FileOperation/bitmap.cpp)
+    ${LIB_DIR}/file/bmp_image.cpp)
 target_link_libraries(example_filters
     PRIVATE example_features_options example_features_warnings)
