+++ examples/filters/makefile
@@ -3,7 +3,7 @@
 ##
 CXXFLAGS += -std=c++11 -Wall -Wextra -Wstrict-aliasing -Wpedantic -Wconversion -O2 -march=native
 
-example_filters : ../../src/image_function_helper.cpp ../../src/image_function.cpp ../../src/filtering.cpp ../../src/FileOperation/bitmap.cpp
+example_filters : ../../src/image_function_helper.cpp ../../src/image_function.cpp ../../src/filtering.cpp ../../src/file/bmp_image.cpp
 
 .PHONY: clean
 clean:
