+++ examples/opencl/image_function/CMakeLists.txt
@@ -4,7 +4,7 @@ add_executable(example_opencl_image_function
     ${LIB_DIR}/opencl/opencl_helper.cpp
     ${LIB_DIR}/image_function.cpp
     ${LIB_DIR}/image_function_helper.cpp
-    ${LIB_DIR}/FileOperation/bitmap.cpp
+    ${LIB_DIR}/file/bmp_image.cpp
     ${LIB_DIR}/opencl/image_function_opencl.cpp)
 target_compile_definitions(example_opencl_image_function
     PRIVATE _CRT_SECURE_NO_WARNINGS)
