+++ examples/windows/edge_detection/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_executable(example_edge_detection
     ${CMAKE_CURRENT_LIST_DIR}/main.cpp
-    ${LIB_DIR}/FileOperation/bitmap.cpp
+    ${LIB_DIR}/file/bmp_image.cpp
     ${LIB_DIR}/ui/ui.cpp
     ${LIB_DIR}/ui/win/win_ui.cpp
     ${LIB_DIR}/image_function_helper.cpp
