+++ examples/windows/edge_detection/example_edge_detection.vcxproj
@@ -149,12 +149,12 @@
     <ClCompile Include="..\..\..\src\ui\ui.cpp" />
     <ClCompile Include="..\..\..\src\ui\win\win_ui.cpp" />
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="..\..\..\src\FileOperation\bitmap.cpp" />
+    <ClCompile Include="..\..\..\src\file\bmp_image.cpp" />
     <ClCompile Include="..\..\..\src\image_function.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\..\src\edge_detection.h" />
-    <ClInclude Include="..\..\..\src\FileOperation\bitmap.h" />
+    <ClInclude Include="..\..\..\src\file\bmp_image.h" />
     <ClInclude Include="..\..\..\src\image_buffer.h" />
     <ClInclude Include="..\..\..\src\image_exception.h" />
     <ClInclude Include="..\..\..\src\image_function.h" />
