+++ examples/windows/edge_detection/main.cpp
@@ -3,7 +3,7 @@
 #include "../../../src/edge_detection.h"
 #include "../../../src/image_buffer.h"
 #include "../../../src/image_function.h"
-#include "../../../src/FileOperation/bitmap.h"
+#include "../../../src/file/bmp_image.h"
 #include "../../../src/ui/win/win_ui.h"
 
 int main()
