+++ src/file/bmp_image.h
