+++ src/file/png_image.h
