+++ src/python/setup.py
@@ -3,7 +3,7 @@ import os
 
 module = Extension( '_penguinV', 
                     sources = ['penguinV.i',
-                               os.path.join('..', 'FileOperation', 'bitmap.cpp'),
+                               os.path.join('..', 'file', 'bmp_image.cpp'),
                                os.path.join('..', 'image_function.cpp'),
                                os.path.join('..', 'image_function_helper.cpp'),  # Need image_function_helper.cpp
                                                                                  # or we get linking errors.
