+++ test/unit_tests/unit_tests.vcxproj
@@ -146,7 +146,7 @@
     <ClCompile Include="..\..\src\blob_detection.cpp" />
     <ClCompile Include="..\..\src\edge_detection.cpp" />
     <ClCompile Include="..\..\src\fft.cpp" />
-    <ClCompile Include="..\..\src\FileOperation\bitmap.cpp" />
+    <ClCompile Include="..\..\src\file\bmp_image.cpp" />
     <ClCompile Include="..\..\src\filtering.cpp" />
     <ClCompile Include="..\..\src\function_pool.cpp" />
     <ClCompile Include="..\..\src\function_pool_task.cpp" />
@@ -175,7 +175,7 @@
     <ClInclude Include="..\..\src\blob_detection.h" />
     <ClInclude Include="..\..\src\edge_detection.h" />
     <ClInclude Include="..\..\src\fft.h" />
-    <ClInclude Include="..\..\src\FileOperation\bitmap.h" />
+    <ClInclude Include="..\..\src\file\bmp_image.h" />
     <ClInclude Include="..\..\src\filtering.h" />
     <ClInclude Include="..\..\src\function_pool.h" />
     <ClInclude Include="..\..\src\function_pool_task.h" />
